requirements.txt
asgiref==3.7.2
coverage==7.4.1
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pytest==8.0.0
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.4
uritemplate==4.1.1
